+++ packages/dash-table/.storybook/webpack.config.js
@@ -1,6 +1,9 @@
 let babel = require('./babel.config.js');
 let config = require('./../.config/webpack/base.js')({
     babel,
+    ts: {
+        transpileOnly: true
+    },
     preprocessor: {
         variables: {
             mode: 'eager'