+++ packages/dash-table/src/dash-table/LazyLoader.ts
@@ -3,6 +3,10 @@ export default class LazyLoader {
         return import(/* webpackChunkName: "export", webpackMode: "$${{mode}}" */ 'xlsx');
     }
 
+    public static get hljs() {
+        return Promise.resolve(window.hljs || import(/* webpackChunkName: "highlight", webpackMode: "$${{mode}}" */ '../third-party/highlight.js').then(result => result.default));
+    }
+
     public static table() {
         return import(/* webpackChunkName: "table", webpackMode: "$${{mode}}" */ 'dash-table/dash/fragments/DataTable');
     }